refactor: UserResponse 응답 객체 구현
