+++ src/main/java/cart/domain/user/UserService.java
@@ -1,8 +1,10 @@
 package cart.domain.user;
 
+import cart.web.controller.user.dto.UserResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class UserService {
@@ -13,7 +15,11 @@ public class UserService {
         this.userDao = userDao;
     }
 
-    public List<User> getUsers() {
-        return userDao.findAll();
+    public List<UserResponse> getUsers() {
+        final List<User> users = userDao.findAll();
+
+        return users.stream()
+                .map(user -> new UserResponse(user.getUserEmailValue(), user.getUserPasswordValue()))
+                .collect(Collectors.toList());
     }
 }
