+++ src/main/java/cart/web/controller/product/ProductViewController.java
@@ -1,9 +1,9 @@
-package cart.web.controller;
+package cart.web.controller.product;
 
 import cart.dao.ProductCategory;
 import cart.domain.product.ProductService;
-import cart.web.controller.dto.ProductRequest;
-import cart.web.controller.dto.ProductResponse;
+import cart.web.controller.product.dto.ProductRequest;
+import cart.web.controller.product.dto.ProductResponse;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
