+++ src/main/java/cart/web/controller/user/UserViewController.java
@@ -1,7 +1,7 @@
-package cart.web.controller;
+package cart.web.controller.user;
 
-import cart.domain.user.User;
 import cart.domain.user.UserService;
+import cart.web.controller.user.dto.UserResponse;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,8 +19,8 @@ public class UserViewController {
 
     @GetMapping("/settings")
     public String renderSettings(final Model model) {
-        final List<User> users = userService.getUsers();
-        model.addAttribute("members", users);
+        final List<UserResponse> userResponses = userService.getUsers();
+        model.addAttribute("members", userResponses);
         return "settings.html";
     }
 }
