+++ src/test/java/cart/service/ProductServiceTest.java
@@ -4,8 +4,8 @@ import cart.dao.ProductCategory;
 import cart.dao.ProductDaoImpl;
 import cart.domain.product.Product;
 import cart.domain.product.ProductService;
-import cart.web.controller.dto.ProductRequest;
-import cart.web.controller.dto.ProductResponse;
+import cart.web.controller.product.dto.ProductRequest;
+import cart.web.controller.product.dto.ProductResponse;
 import cart.web.exception.GlobalException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
